chore: lazy load routes
