+++ admin/src/pages/collections/[name]/edit.tsx
@@ -7,7 +7,7 @@ import Collection from '../../../client/collection'
 import CollectionForm from '../../../components/collection-form'
 import { CollectionRouteData } from '../../../routes'
 
-function Edit() {
+function Component() {
   const { collection } = useRouteLoaderData('collection') as CollectionRouteData
   const revalidator = useRevalidator()
 
@@ -31,4 +31,4 @@ function Edit() {
   )
 }
 
-export default Edit
+export { Component }
